requirements.txt
django
djangorestframework
django-admin-interface
django-cors-headers
django-import-export
python-dotenv